(FEAN-222) feat : Coupon 최소 주문금액, 최대 할인금액 Double -> Long 수정

+++ src/main/java/com/lotte/danuri/product/model/dto/CouponDto.java
@@ -24,8 +24,8 @@ public class CouponDto {
     private LocalDateTime endDate;
 
     private Double discountRate;
-    private Double minOrderPrice;
-    private Double maxDiscountPrice;
+    private Long minOrderPrice;
+    private Long maxDiscountPrice;
 
     private List<Long> productId;
